NEW: TrialIndexer
- can be used both as an iterable or via explicit (single element) indexing

Changes to be committed:
	modified:   syncopy/__init__.py
	modified:   syncopy/datatype/__init__.py
	modified:   syncopy/datatype/base_data.py
	modified:   syncopy/datatype/discrete_data.py
	new file:   syncopy/datatype/util.py
	modified:   syncopy/statistics/spike_psth.py
	modified:   syncopy/statistics/summary_stats.py
	modified:   syncopy/tests/test_selectdata.py
