+++ syncopy/datatype/__init__.py
@@ -13,6 +13,7 @@ from .methods.padding import *
 from .methods.selectdata import *
 from .methods.show import *
 from .methods.copy import *
+from .util import *
 
 # Populate local __all__ namespace
 __all__ = []
@@ -25,3 +26,4 @@ __all__.extend(methods.definetrial.__all__)
 __all__.extend(methods.selectdata.__all__)
 __all__.extend(methods.show.__all__)
 __all__.extend(methods.copy.__all__)
+__all__.extend(util.__all__)
